fix: address billing reconciliation review follow-ups
