+++ web/default/src/features/smart-ops/components/billing-settlement-evidence.tsx
@@ -17,15 +17,10 @@ along with this program. If not, see <https://www.gnu.org/licenses/>.
 For commercial licensing, please contact https://github.com/MAX-API-Next/MAX-API/issues
 */
 import { useState, type ReactElement } from 'react'
-import type { TFunction } from 'i18next'
-import { CheckCircle2, Pencil, TriangleAlert } from 'lucide-react'
+import { CheckCircle2, TriangleAlert } from 'lucide-react'
 import { useTranslation } from 'react-i18next'
 import { toast } from 'sonner'
-import {
-  formatQuota,
-  formatTimestampRelative,
-  formatTimestampToDate,
-} from '@/lib/format'
+import { formatTimestampToDate } from '@/lib/format'
 import { handleServerError } from '@/lib/handle-server-error'
 import { Alert, AlertDescription, AlertTitle } from '@/components/ui/alert'
 import { Badge } from '@/components/ui/badge'
@@ -38,14 +33,6 @@ import {
 } from '@/components/ui/field'
 import { Skeleton } from '@/components/ui/skeleton'
 import { Switch } from '@/components/ui/switch'
-import {
-  Table,
-  TableBody,
-  TableCell,
-  TableHead,
-  TableHeader,
-  TableRow,
-} from '@/components/ui/table'
 import { updateBillingSettlementBlockingPolicy } from '../api'
 import { formatCount, formatLocalizedCount } from '../lib/format'
 import { mutationErrorMessage } from '../lib/mutation-error'
@@ -53,29 +40,8 @@ import type {
   BillingSettlementReconciliationData,
   BillingSettlementReconciliationItem,
 } from '../types'
+import { BillingSettlementTable } from './billing-settlement-table'
 import { SettlementReviewDialog } from './settlement-review-dialog'
-
-const FUNDING_SOURCE_LABEL: Record<string, string> = {
-  wallet: 'Wallet',
-  subscription: 'Subscription',
-}
-
-function settlementReferences(
-  item: BillingSettlementReconciliationItem,
-  t: TFunction
-): string[] {
-  const references = [t('User #{{id}}', { id: item.user_id })]
-  if (item.subscription_id > 0) {
-    references.push(t('Subscription #{{id}}', { id: item.subscription_id }))
-  }
-  if (item.token_id > 0) {
-    references.push(t('Token #{{id}}', { id: item.token_id }))
-  }
-  if (item.task_id > 0) {
-    references.push(t('Task #{{id}}', { id: item.task_id }))
-  }
-  return references
-}
 
 interface BillingSettlementEvidenceProps {
   canUpdateBlockingPolicy: boolean
@@ -193,7 +159,7 @@ export function BillingSettlementEvidence(
               i18n.language,
               t,
               'Reviewed record: {{count}}',
-              'Reviewed: {{count}}'
+              'Reviewed records: {{count}}'
             )}
           </Badge>
           <Badge variant='outline'>
@@ -254,189 +220,10 @@ export function BillingSettlementEvidence(
           </AlertDescription>
         </Alert>
       ) : (
-        <div className='overflow-x-auto rounded-md border'>
-          <Table className='min-w-[1580px]'>
-            <TableHeader>
-              <TableRow className='bg-muted/40 hover:bg-muted/40'>
-                <TableHead>{t('Operation')}</TableHead>
-                <TableHead>{t('Financial / alert status')}</TableHead>
-                <TableHead>{t('User access')}</TableHead>
-                <TableHead>{t('References')}</TableHead>
-                <TableHead>{t('Funding source')}</TableHead>
-                <TableHead className='text-right'>
-                  {t('Outstanding funding')}
-                </TableHead>
-                <TableHead>{t('Attempts / next retry')}</TableHead>
-                <TableHead>{t('Last error')}</TableHead>
-                <TableHead>{t('Administrator review')}</TableHead>
-                <TableHead>{t('Created')}</TableHead>
-                <TableHead className='text-right'>{t('Actions')}</TableHead>
-              </TableRow>
-            </TableHeader>
-            <TableBody>
-              {props.data.items.map((item) => {
-                const outstandingFunding =
-                  item.funding_delta - item.applied_funding_delta
-                const references = settlementReferences(item, t)
-                return (
-                  <TableRow key={item.id}>
-                    <TableCell className='max-w-72 whitespace-normal'>
-                      <span className='font-mono text-xs break-all'>
-                        {item.operation_key}
-                      </span>
-                    </TableCell>
-                    <TableCell>
-                      <div className='flex flex-col items-start gap-1'>
-                        <Badge
-                          variant={
-                            item.status === 'manual' ? 'destructive' : 'outline'
-                          }
-                        >
-                          {item.status === 'manual'
-                            ? t('Manual')
-                            : t('Pending')}
-                        </Badge>
-                        <Badge
-                          variant={
-                            item.reconciliation_reviewed_at > 0
-                              ? 'secondary'
-                              : 'destructive'
-                          }
-                        >
-                          {item.reconciliation_reviewed_at > 0
-                            ? t('Reviewed')
-                            : t('Open alert')}
-                        </Badge>
-                      </div>
-                    </TableCell>
-                    <TableCell>
-                      <div className='flex flex-col items-start gap-1'>
-                        <Badge
-                          variant={
-                            item.blocks_user ? 'destructive' : 'secondary'
-                          }
-                        >
-                          {item.blocks_user
-                            ? t('User blocked')
-                            : t('User allowed')}
-                        </Badge>
-                        <span className='text-muted-foreground text-xs'>
-                          {item.record_blocks_user
-                            ? t('Record policy: block')
-                            : t('Record policy: allow')}
-                        </span>
-                      </div>
-                    </TableCell>
-                    <TableCell className='whitespace-normal'>
-                      <div className='flex max-w-48 flex-col gap-0.5 font-mono text-xs'>
-                        {references.map((reference) => (
-                          <span key={reference}>{reference}</span>
-                        ))}
-                      </div>
-                    </TableCell>
-                    <TableCell>
-                      {t(FUNDING_SOURCE_LABEL[item.source] ?? item.source)}
-                    </TableCell>
-                    <TableCell className='text-right font-mono tabular-nums'>
-                      {formatQuota(outstandingFunding)}
-                    </TableCell>
-                    <TableCell className='whitespace-normal'>
-                      <div className='flex max-w-44 flex-col gap-0.5 text-xs'>
-                        <span>
-                          {formatLocalizedCount(
-                            item.attempts,
-                            i18n.language,
-                            t,
-                            '{{count}} attempt',
-                            '{{count}} attempts'
-                          )}
-                        </span>
-                        <span
-                          className='text-muted-foreground'
-                          title={formatTimestampToDate(item.next_attempt)}
-                        >
-                          {item.status === 'manual'
-                            ? t('Manual review')
-                            : formatTimestampRelative(
-                                item.next_attempt,
-                                'seconds',
-                                i18n.language
-                              )}
-                        </span>
-                      </div>
-                    </TableCell>
-                    <TableCell className='max-w-72 whitespace-normal'>
-                      <span
-                        className='text-muted-foreground text-xs break-words'
-                        title={item.last_error || undefined}
-                      >
-                        {item.last_error || t('No error detail')}
-                      </span>
-                    </TableCell>
-                    <TableCell className='max-w-72 whitespace-normal'>
-                      {item.reconciliation_reviewed_at > 0 ? (
-                        <div className='flex flex-col gap-0.5 text-xs'>
-                          <span>
-                            {t('Administrator #{{id}}', {
-                              id: item.reconciliation_reviewed_by,
-                            })}
-                          </span>
-                          <span
-                            className='text-muted-foreground'
-                            title={formatTimestampToDate(
-                              item.reconciliation_reviewed_at
-                            )}
-                          >
-                            {formatTimestampRelative(
-                              item.reconciliation_reviewed_at,
-                              'seconds',
-                              i18n.language
-                            )}
-                          </span>
-                          <span
-                            className='text-muted-foreground line-clamp-3'
-                            title={item.reconciliation_review_note}
-                          >
-                            {item.reconciliation_review_note}
-                          </span>
-                        </div>
-                      ) : (
-                        <span className='text-muted-foreground text-xs'>
-                          {t('Not reviewed')}
-                        </span>
-                      )}
-                    </TableCell>
-                    <TableCell
-                      className='text-muted-foreground text-xs'
-                      title={formatTimestampToDate(item.created_at)}
-                    >
-                      {formatTimestampRelative(
-                        item.created_at,
-                        'seconds',
-                        i18n.language
-                      )}
-                    </TableCell>
-                    <TableCell className='text-right'>
-                      <Button
-                        type='button'
-                        variant='outline'
-                        size='sm'
-                        onClick={() => setSelectedItem(item)}
-                      >
-                        {item.reconciliation_reviewed_at > 0 && (
-                          <Pencil data-icon='inline-start' aria-hidden='true' />
-                        )}
-                        {item.reconciliation_reviewed_at > 0
-                          ? t('Edit review')
-                          : t('Review and close')}
-                      </Button>
-                    </TableCell>
-                  </TableRow>
-                )
-              })}
-            </TableBody>
-          </Table>
-        </div>
+        <BillingSettlementTable
+          items={props.data.items}
+          onSelectItem={setSelectedItem}
+        />
       )}
 
       {props.data.truncated && (
