+++ web/default/src/i18n/locales/fr.json
@@ -5554,8 +5554,8 @@
     "Review note": "Note de vérification",
     "Review note must contain between 3 and 1000 characters.": "La note de vérification doit contenir entre 3 et 1000 caractères.",
     "Reviewed": "Vérifié",
-    "Reviewed record: {{count}}": "Vérifié : {{count}}",
-    "Reviewed: {{count}}": "Vérifiés : {{count}}",
+    "Reviewed record: {{count}}": "Enregistrement vérifié : {{count}}",
+    "Reviewed records: {{count}}": "Enregistrements vérifiés : {{count}}",
     "Save review": "Enregistrer la vérification",
     "There are no pending or manual positive final settlements requiring operator review.": "Aucun règlement final positif en attente ou manuel ne nécessite de vérification opérateur.",
     "This choice affects only new paid-request admission while the financial settlement remains pending or manual.": "Ce choix affecte uniquement l’admission de nouvelles requêtes payantes tant que le règlement financier reste en attente ou manuel.",
