+++ web/default/src/i18n/locales/ja.json
@@ -5554,8 +5554,8 @@
     "Review note": "確認メモ",
     "Review note must contain between 3 and 1000 characters.": "確認メモは 3～1000 文字で入力してください。",
     "Reviewed": "確認済み",
-    "Reviewed record: {{count}}": "確認済み：{{count}}",
-    "Reviewed: {{count}}": "確認済み：{{count}}",
+    "Reviewed record: {{count}}": "確認済みレコード：{{count}}",
+    "Reviewed records: {{count}}": "確認済みレコード：{{count}}",
     "Save review": "確認を保存",
     "There are no pending or manual positive final settlements requiring operator review.": "管理者による確認が必要な pending または manual の正の最終決済はありません。",
     "This choice affects only new paid-request admission while the financial settlement remains pending or manual.": "この選択は、財務決済が pending または manual の間における新規有料リクエストの受付だけに影響します。",
