+++ web/default/src/i18n/locales/ru.json
@@ -3,7 +3,7 @@
     "360": "360",
     "1000": "1000",
     "10000": "10000",
-    "{{count}} attempt": "{{count}} попытка",
+    "{{count}} attempt": "Попыток: {{count}}",
     "{{count}} attempts": "Попыток: {{count}}",
     "{{count}} day": "{{count}} день",
     "{{count}} days": "{{count}} дней",
@@ -11,7 +11,7 @@
     "{{count}} hours": "{{count}} часов",
     "{{count}} minute": "{{count}} минута",
     "{{count}} minutes": "{{count}} минут",
-    "{{count}} record": "Запись: {{count}}",
+    "{{count}} record": "Записей: {{count}}",
     "{{count}} records": "Записей: {{count}}",
     "{{count}} second": "{{count}} секунда",
     "{{count}} seconds": "{{count}} секунд",
@@ -5555,7 +5555,7 @@
     "Review note must contain between 3 and 1000 characters.": "Примечание к проверке должно содержать от 3 до 1000 символов.",
     "Reviewed": "Проверено",
     "Reviewed record: {{count}}": "Проверена запись: {{count}}",
-    "Reviewed: {{count}}": "Проверено: {{count}}",
+    "Reviewed records: {{count}}": "Проверено записей: {{count}}",
     "Save review": "Сохранить проверку",
     "There are no pending or manual positive final settlements requiring operator review.": "Нет положительных окончательных расчётов в состоянии pending или manual, требующих проверки оператором.",
     "This choice affects only new paid-request admission while the financial settlement remains pending or manual.": "Этот выбор влияет только на допуск новых платных запросов, пока финансовый расчёт остаётся в состоянии pending или manual.",
