+++ web/default/src/i18n/locales/vi.json
@@ -5549,13 +5549,13 @@
     "Record policy: block": "Chính sách bản ghi: chặn",
     "Record the evidence checked, the conclusion, and any follow-up action.": "Ghi lại bằng chứng đã kiểm tra, kết luận và hành động tiếp theo.",
     "Review and close": "Kiểm tra và đóng",
-    "Review and close operational alerts without changing the underlying pending or manual financial settlement.": "Kiểm tra và đóng cảnh báo vận hành mà không thay đổi quyết toán tài chính đang chờ hoặc xử lý thủ công bên dưới.",
+    "Review and close operational alerts without changing the underlying pending or manual financial settlement.": "Kiểm tra và đóng cảnh báo vận hành mà không thay đổi quyết toán tài chính tương ứng đang chờ hoặc xử lý thủ công.",
     "Review billing reconciliation": "Kiểm tra đối soát tính phí",
     "Review note": "Ghi chú kiểm tra",
     "Review note must contain between 3 and 1000 characters.": "Ghi chú kiểm tra phải có từ 3 đến 1000 ký tự.",
     "Reviewed": "Đã kiểm tra",
-    "Reviewed record: {{count}}": "Đã kiểm tra: {{count}}",
-    "Reviewed: {{count}}": "Đã kiểm tra: {{count}}",
+    "Reviewed record: {{count}}": "Đã kiểm tra bản ghi: {{count}}",
+    "Reviewed records: {{count}}": "Đã kiểm tra bản ghi: {{count}}",
     "Save review": "Lưu kết quả kiểm tra",
     "There are no pending or manual positive final settlements requiring operator review.": "Không có quyết toán cuối dương đang chờ hoặc xử lý thủ công nào cần quản trị viên kiểm tra.",
     "This choice affects only new paid-request admission while the financial settlement remains pending or manual.": "Lựa chọn này chỉ ảnh hưởng đến việc tiếp nhận yêu cầu trả phí mới trong khi quyết toán tài chính vẫn ở trạng thái đang chờ hoặc xử lý thủ công.",
