+++ web/default/src/i18n/locales/zh.json
@@ -5541,7 +5541,7 @@
     "No unresolved reconciliation records.": "没有未完成的对账记录。",
     "Node": "节点",
     "Not reviewed": "未核对",
-    "Only root administrators can change the default blocking policy.": "仅超级管理员可更改默认阻塞策略。",
+    "Only root administrators can change the default blocking policy.": "仅 Root 管理员可更改默认阻塞策略。",
     "Open alert": "打开告警",
     "Open alert: {{count}}": "未关闭告警：{{count}}",
     "Open alerts: {{count}}": "未关闭告警：{{count}}",
@@ -5554,8 +5554,8 @@
     "Review note": "核对备注",
     "Review note must contain between 3 and 1000 characters.": "核对备注必须包含 3 到 1000 个字符。",
     "Reviewed": "已核对",
-    "Reviewed record: {{count}}": "已核对：{{count}}",
-    "Reviewed: {{count}}": "已核对：{{count}}",
+    "Reviewed record: {{count}}": "已核对记录：{{count}}",
+    "Reviewed records: {{count}}": "已核对记录：{{count}}",
     "Save review": "保存核对",
     "There are no pending or manual positive final settlements requiring operator review.": "没有需要管理员核对的待处理或人工处理正向最终结算。",
     "This choice affects only new paid-request admission while the financial settlement remains pending or manual.": "此选择只影响财务结算保持待处理或人工处理期间的新付费请求准入。",
